requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0
scipy>=1.10
numpy>=1.24
tqdm>=4.65